requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.4.1
django-cors-headers==4.6.0
django-websocket-redis==0.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
firebase-admin==6.6.0
frozenlist==1.5.0
gevent==24.11.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
twilio==9.4.4
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
